refactor: Decouple HF Hub utils and spaCy utils from model_loading module
